refactor: moved fetch_repos function from separate script to util.py
